# use consecutive WriteString calls instead of Fprintf