It seems like it's faster to use those calls than invoke Fprintf:

```
benchmark                        old ns/op     new ns/op     delta
BenchmarkStreamNoCtx-4           4755          3937          -17.20%
BenchmarkDiscard-4               816           787           -3.55%
BenchmarkCallerFileHandler-4     1879          1811          -3.62%
BenchmarkCallerFuncHandler-4     1656          1663          +0.42%
BenchmarkLogfmtNoCtx-4           3556          2784          -21.71%
BenchmarkJsonNoCtx-4             1599          1614          +0.94%
BenchmarkMultiLevelFilter-4      862           833           -3.36%
BenchmarkDescendant1-4           839           813           -3.10%
BenchmarkDescendant2-4           846           820           -3.07%
BenchmarkDescendant4-4           918           897           -2.29%
BenchmarkDescendant8-4           971           964           -0.72%
```

Fixes some places in the tests that panic if the tests don't exit immediately.